# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: ZVideo API
  version: 1.0.0
  description: REST API for a video hosting platform.
servers:
  - url: /api/v1
security:
  - BearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotate refresh token and issue new tokens
      operationId: refreshTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Tokens refreshed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /channels:
    get:
      tags: [Channels]
      security: []
      summary: List channels
      operationId: listChannels
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Channels page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelPage'
    post:
      tags: [Channels]
      summary: Create current user's channel
      operationId: createChannel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelCreateRequest'
      responses:
        '201':
          description: Channel created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
  /channels/me:
    get:
      tags: [Channels]
      summary: Get current user's channel
      operationId: getMyChannel
      responses:
        '200':
          description: Current user's channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}:
    get:
      tags: [Channels]
      security: []
      summary: Get channel by id
      operationId: getChannel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Channels]
      summary: Update channel
      operationId: updateChannel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChannelUpdateRequest'
      responses:
        '200':
          description: Channel updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    delete:
      tags: [Channels]
      summary: Delete channel
      operationId: deleteChannel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '204':
          description: Channel deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/by-name/{name}:
    get:
      tags: [Channels]
      security: []
      summary: Get channel by name
      operationId: getChannelByName
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Channel.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Channel'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/subscriptions:
    put:
      tags: [Subscriptions]
      summary: Subscribe to channel
      operationId: subscribeToChannel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '204':
          description: Subscribed.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Subscriptions]
      summary: Unsubscribe from channel
      operationId: unsubscribeFromChannel
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '204':
          description: Unsubscribed.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/subscriptions/status:
    get:
      tags: [Subscriptions]
      summary: Check current user's subscription to channel
      operationId: getChannelSubscriptionStatus
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Subscription status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionStatus'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /channels/{channelId}/subscribers/count:
    get:
      tags: [Subscriptions]
      security: []
      summary: Get channel subscribers count
      operationId: getChannelSubscribersCount
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '200':
          description: Subscribers count.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CountResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/me/subscriptions:
    get:
      tags: [Subscriptions]
      summary: List current user's subscriptions
      operationId: listMySubscriptions
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: Subscriptions page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SubscriptionPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /users/me/subscriptions/{channelId}/new-videos-count:
    delete:
      tags: [Subscriptions]
      summary: Reset unread new videos count for subscription
      operationId: resetSubscriptionNewVideosCount
      parameters:
        - $ref: '#/components/parameters/ChannelId'
      responses:
        '204':
          description: Counter reset.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    ChannelId:
      name: channelId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    BadRequest:
      description: Request validation or domain rule failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidUsername:
              summary: Invalid username
              value:
                error:
                  code: INVALID_USERNAME
                  message: The username is invalid
            invalidUserEmail:
              summary: Invalid email
              value:
                error:
                  code: INVALID_USER_EMAIL
                  message: The email is invalid
            weakPassword:
              summary: Weak password
              value:
                error:
                  code: WEAK_PASSWORD
                  message: The password is too weak
            playlistNameEmpty:
              summary: Empty playlist name
              value:
                error:
                  code: PLAYLIST_NAME_EMPTY
                  message: Playlist name cannot be empty
            playlistVideoChannelMismatch:
              summary: Playlist and video belong to different channels
              value:
                error:
                  code: PLAYLIST_VIDEO_CHANNEL_MISMATCH
                  message: Video belongs to another channel
            selfSubscription:
              summary: User cannot subscribe to own channel
              value:
                error:
                  code: SELF_SUBSCRIPTION
                  message: Cannot subscribe to your own channel
            invalidNotificationSettings:
              summary: Invalid notification settings
              value:
                error:
                  code: INVALID_NOTIFICATION_SETTINGS
                  message: Notification settings payload is invalid
            roleNotFound:
              summary: Role not found
              value:
                error:
                  code: ROLE_NOT_FOUND
                  message: Role not found
            communityPostContentEmpty:
              summary: Empty community post
              value:
                error:
                  code: COMMUNITY_POST_CONTENT_EMPTY
                  message: Community post content cannot be empty
            communityCommentContentEmpty:
              summary: Empty community comment
              value:
                error:
                  code: COMMUNITY_COMMENT_CONTENT_EMPTY
                  message: Community comment content cannot be empty
    Unauthorized:
      description: Authentication failed or token is missing.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidCredentials:
              summary: Invalid login credentials
              value:
                error:
                  code: INVALID_CREDENTIALS
                  message: Invalid email or password
            invalidToken:
              summary: Invalid or expired token
              value:
                error:
                  code: INVALID_TOKEN
                  message: Invalid or expired token
            missingAuthorization:
              summary: Missing bearer token
              value:
                error:
                  code: UNAUTHORIZED
                  message: user not authorized
    Forbidden:
      description: Authenticated user is not allowed to perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            forbidden:
              summary: Insufficient permissions
              value:
                error:
                  code: FORBIDDEN
                  message: You don't have permission to access this resource
            userBanned:
              summary: User account is banned
              value:
                error:
                  code: USER_BANNED
                  message: User account is banned
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            userNotFound:
              summary: User not found
              value:
                error:
                  code: USER_NOT_FOUND
                  message: User not found
            channelNotFound:
              summary: Channel not found
              value:
                error:
                  code: CHANNEL_NOT_FOUND
                  message: Channel not found
            videoNotFound:
              summary: Video not found
              value:
                error:
                  code: VIDEO_NOT_FOUND
                  message: Video not found
            playlistNotFound:
              summary: Playlist not found
              value:
                error:
                  code: PLAYLIST_NOT_FOUND
                  message: Playlist not found
            ratingNotFound:
              summary: Rating not found
              value:
                error:
                  code: RATING_NOT_FOUND
                  message: Rating not found
            commentNotFound:
              summary: Comment not found
              value:
                error:
                  code: COMMENT_NOT_FOUND
                  message: Comment not found
            communityPostNotFound:
              summary: Community post not found
              value:
                error:
                  code: COMMUNITY_POST_NOT_FOUND
                  message: Community post not found
            communityCommentNotFound:
              summary: Community comment not found
              value:
                error:
                  code: COMMUNITY_COMMENT_NOT_FOUND
                  message: Community comment not found
    Conflict:
      description: Resource state conflicts with the request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            usernameTaken:
              summary: Username already exists
              value:
                error:
                  code: USERNAME_TAKEN
                  message: Username already taken
            emailTaken:
              summary: Email already exists
              value:
                error:
                  code: EMAIL_TAKEN
                  message: Email already registered
            channelAlreadyExists:
              summary: User already has a channel
              value:
                error:
                  code: CHANNEL_ALREADY_EXISTS
                  message: Channel already exists
            channelNameTaken:
              summary: Channel name already exists
              value:
                error:
                  code: CHANNEL_NAME_TAKEN
                  message: Channel name already exists
            alreadyRated:
              summary: Resource has already been rated
              value:
                error:
                  code: ALREADY_RATED
                  message: This video has already been rated
            videoNotReady:
              summary: Video is not ready
              value:
                error:
                  code: VIDEO_NOT_READY
                  message: Video is not ready yet
    InternalServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            internalError:
              summary: Unexpected server error
              value:
                error:
                  code: INTERNAL_ERROR
                  message: Something went wrong
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Error'
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: ERROR_CODE
        message:
          type: string
          example: Human-readable error message
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 8
...
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    AuthResponse:
      type: object
      required: [user, accessToken, refreshToken, refreshExpiresAt]
      properties:
        user:
          $ref: '#/components/schemas/User'
        accessToken:
          type: string
        refreshToken:
          type: string
        refreshExpiresAt:
          type: string
          format: date-time
    Role:
      type: object
      required: [id, name, isDefault]
      properties:
        id:
          type: integer
        name:
          $ref: '#/components/schemas/RoleName'
        isDefault:
          type: boolean
    RoleName:
      type: string
      enum: [admin, moderator, user]
    User:
      type: object
      required: [id, username, email, isActive, notificationsEnabled, createdAt, updatedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        isActive:
          type: boolean
        notificationsEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        role:
          $ref: '#/components/schemas/Role'
...
    Channel:
      type: object
      required: [id, userId, ownerUsername, name, description, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        ownerUsername:
          type: string
        name:
          type: string
        description:
          type: string
        createdAt:
          type: string
          format: date-time
    ChannelCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
          default: ''
    ChannelUpdateRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 32
        description:
          type: string
    Subscription:
      type: object
      required: [userId, channelId, channelName, newVideosCount, subscribedAt]
      properties:
        userId:
          type: integer
        channelId:
          type: integer
        channelName:
          type: string
        newVideosCount:
          type: integer
          minimum: 0
        subscribedAt:
          type: string
          format: date-time
    SubscriptionStatus:
      type: object
      required: [subscribed]
      properties:
        subscribed:
          type: boolean
    CountResponse:
      type: object
      required: [count]
      properties:
        count:
          type: integer
          minimum: 0
...
    PageMeta:
      type: object
      required: [limit, offset, totalCount]
      properties:
        limit:
          type: integer
        offset:
          type: integer
        totalCount:
          type: integer
          format: int64
          minimum: 0
...
    ChannelPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Channel'
    SubscriptionPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/Subscription'
...